% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-addons.r
\name{label_facet}
\alias{guess_dim_facet}
\alias{label_facet}
\title{Label facets}
\usage{
label_facet(facets, data, labels = paste0(LETTERS, ")"), x = -Inf,
  y = Inf, hjust = -0.2, vjust = 1.4, size = rel(3.6),
  fontface = "bold", ...)

guess_dim_facet(facets, data)
}
\arguments{
\item{facets}{A \code{formula}: the same that is supplied to \code{facet_grid}}

\item{data}{a \code{data.frame}:}

\item{labels}{labels to plot}

\item{x}{see \code{?geom_text}}

\item{y}{see \code{?geom_text}}

\item{hjust}{see \code{?geom_text}}

\item{vjust}{see \code{?geom_text}}

\item{size}{see \code{?geom_text}}

\item{fontface}{see \code{?geom_text}}

\item{...}{extra arguments to \code{geom_text}}
}
\value{
\code{geom_text} object
}
\description{
A convenience function for labeling facets for publication
}
\details{
needs the same \code{data.frame} and facet formula as the plot on top of which it is being layered
}

