% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-addons.r
\name{label_parser}
\alias{label_parser}
\title{A more robust label parser}
\usage{
label_parser(variable, value)
}
\arguments{
\item{variable}{variable (ggplot internal)}

\item{value}{variable (ggplot internal)}
}
\value{
a \code{list} that is useful for panel text (i.e. \code{strip.text.x})
}
\description{
a fillin replacement for \code{label_parsed}
}

