% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-addons.r
\name{print.crap}
\alias{print.crap}
\title{S3 method to plot \code{grid} object result from \code{move_facet_strip_y}}
\usage{
\method{print}{crap}(g)
}
\arguments{
\item{g}{Result from \code{move_facet_strip_y}}
}
\description{
S3 method to plot \code{grid} object result from \code{move_facet_strip_y}
}

