% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-stuff.r
\name{reclassify}
\alias{reclassify}
\title{Rename characters or factors based on their name, as determined by the supplied  regular expression.}
\usage{
reclassify(stuff, named)
}
\arguments{
\item{stuff}{character vector to be renamed}

\item{named}{named character vector or list. Syntax is: "to"="from"}
}
\value{
a character vector
}
\description{
Inspired by "revalue" from plyr package
}

