% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename-stuff.r
\name{rerevalue}
\alias{rerevalue}
\title{Rename characters or factors based on their name, as determined by the supplied  regular expression.}
\usage{
rerevalue(x, replace = NULL, warn_missing = TRUE, ...)
}
\arguments{
\item{named}{named character vector or list. Syntax is: "to"="from"}

\item{stuff}{character vector to be renamed}
}
\value{
a character vector
}
\description{
Inspired by "revalue" from plyr package
}

