% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-addons.r
\name{move_facet_strip_y}
\alias{move_facet_strip_x}
\alias{move_facet_strip_y}
\title{Move facet strip y}
\usage{
move_facet_strip_y(object, sane.defaults = TRUE)

move_facet_strip_x(object, sane.defaults = TRUE)
}
\value{
\code{grid} object
}
\description{
A brutish fix to allow plotting of time series stacked on top of each
other.  It will work for more than just time series.  Probably violates a few principles
of the grammar of graphics and usually requires the violation of tidy data
principles.
}
\details{
takes \code{ggplot} object and changes it to a \code{grob}, mauls it a bit, and
    spits out a \code{grob}
From: http://stackoverflow.com/questions/23497682/alter-facet-grid-output-in-ggplot2
}
\examples{
library(reshape2)
## simulate a model study
air <- transform(airquality, Ozone.model = Ozone +
    rnorm(length(Ozone), 2),
                        Wind.model = Wind + rnorm(length(Wind)))
## Add time, see ?airquality
air$dtm <- as.Date(paste("1973", air$Month, air$Day,
                                sep = "-"), format="\%Y-\%m-\%d", tz="EST")
d.f <- melt(air, id.vars = "dtm", measure.vars = c("Wind",
    "Wind.model", "Ozone", "Ozone.model"))
d.f$type <- ifelse(grepl("model", d.f$variable), "Modeled", "Measured")
d.f$Variables <- sub("\\\\.model", "", d.f$variable)
(box <- ggplot(d.f, aes(dtm, value, color=type)) +
    geom_line() +
     facet_grid(Variables~., scales="free_y"))
## please note that \\code{ggsave} will no longer work with earlier  versions (<= 2.0.0) of ggplot2
x11()
move_facet_strip_y(box)

}

